backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
xgboost==2.1.3
shap==0.46.0
scikit-learn==1.6.0
numpy==1.26.4
pandas==2.2.3
Pillow==11.0.0
python-dotenv==1.0.1
